+++ lib/search_field.dart
@@ -1,2 +1,3 @@
 export 'src/search_field.dart';
 export 'src/search_field_data_model.dart';
+export 'src/search_field_controller.dart';