+++ lib/src/Utils.dart
@@ -1,4 +1,5 @@
 class Utils {
+  /// this function is going to build and return key from [text]
   static String buildKey(String text){
     return text.replaceAll(" ", "_");
   }