+++ lib/src/search_field_controller.dart
@@ -3,8 +3,7 @@ import 'package:search_field/search_field.dart';
 class SearchFieldController {
   String text;
   bool isLoading;
-  bool selected;
-  void Function(SearchFieldDataModel item)? fetch;
-  SearchFieldController({this.text ="", this.isLoading = false, this.selected = true});
+  void Function(SearchFieldDataModel model)? selected;
+  SearchFieldController({this.text ="", this.isLoading = false, this.selected});
 }