+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
-
-* TODO: Describe initial release.
+initial release
+* predefined list search/ filter
+* network search query and show suggestion
+* other searchFiled dependency
+* fetch content on dependency search resoved
+* select item from suggestion